// src/types/partials/MeasureGroup.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_classes so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TMeasurePopulation } from '../partials/MeasurePopulation';
import { TMeasureStratifier } from '../partials/MeasureStratifier';

export type TMeasureGroup = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    code?: TCodeableConcept;
    description?: String;
    population?: TMeasurePopulation[];
    stratifier?: TMeasureStratifier[];
};